Price modified to floating point

+++ PaperTrading_backend/trading/models.py
@@ -6,7 +6,7 @@ class Buy(models.Model):
     username = models.CharField(max_length=100)
     user_id = models.CharField(max_length=100)
     stock_name = models.CharField(max_length=100)
-    stock_price = models.IntegerField()  #total price of all stock = each_stock_price * stock_quantity
+    stock_price = models.FloatField()  #total price of all stock = each_stock_price * stock_quantity
     stock_quantity = models.IntegerField()
 
 class Transaction(models.Model):
@@ -14,7 +14,7 @@ class Transaction(models.Model):
     user_id = models.CharField(max_length=100)
     stock_name = models.CharField(max_length=100)
     stock_quantity = models.IntegerField()
-    stock_price = models.IntegerField()
+    stock_price = models.FloatField()
     date_time = models.CharField(max_length=100)
     buy_sell = models.CharField(max_length=100, null=True)
 
@@ -22,9 +22,9 @@ class Bookmark(models.Model):
     username = models.CharField(max_length=100)
     user_id = models.CharField(max_length=100)
     stock_name = models.CharField(max_length=100)
-    stock_price = models.IntegerField()
+    stock_price = models.FloatField()
 
 class Balance(models.Model):
     username = models.CharField(max_length=100)
     user_id = models.CharField(max_length=100)
-    balance = models.IntegerField()
+    balance = models.FloatField()
